+++ imagery/i.atcorr/aerosolconcentration.h
@@ -32,17 +32,17 @@ struct AtmosModel;
 struct AerosolConcentration
 {
 	/* aerosol concentration parameters */
-    float taer55;
+    double taer55;
 
 private:
     long int iaer;
-    float v;
+    double v;
     void parse(const long int iaer, const AtmosModel &atms);
-    void oda550(const float v, const AtmosModel &atms);
+    void oda550(const double v, const AtmosModel &atms);
 
 public:
     /* Set the visibility, this will overide any previous estimates of taer55 */
-    void set_visibility (const float vis, const AtmosModel &atms) { if(vis > 0) oda550(vis, atms); }
+    void set_visibility (const double vis, const AtmosModel &atms) { if(vis > 0) oda550(vis, atms); }
     void print();
     static AerosolConcentration Parse(const long int iaer, const AtmosModel &atms);
 };
